Omaha #3410 changed lon/lat fields to floats in sfc and air obs location

Former-commit-id: 74a6c3dd20898e9bd4484376ea4e4e471bdec9af [formerly a7d17bd490f2f0931f3c22ee5484214041a34bd6 [formerly aa991c922354f0a67fac08fee456ca60a30de968] [formerly 74a6c3dd20898e9bd4484376ea4e4e471bdec9af [formerly cdb0a72fcb2345564f2b944fdecd7869fd498767]]]
Former-commit-id: a7d17bd490f2f0931f3c22ee5484214041a34bd6 [formerly aa991c922354f0a67fac08fee456ca60a30de968]
Former-commit-id: a7d17bd490f2f0931f3c22ee5484214041a34bd6
Former-commit-id: 26ade7cae88d84ef9e1aed95b3df2cd6d5f893ee

+++ edexOsgi/com.raytheon.uf.edex.plugin.acars/src/com/raytheon/uf/edex/plugin/acars/decoder/ACARSDataAdapter.java
@@ -46,9 +46,9 @@ import com.raytheon.uf.edex.decodertools.bufr.packets.IBUFRDataPacket;
 
 /**
  * Adapter used to decode ACARS data in BUFR format.
- *
+ * 
  * <pre>
- *
+ * 
  * SOFTWARE HISTORY
  * Date          Ticket#  Engineer    Description
  * ------------- -------- ----------- --------------------------
@@ -59,9 +59,10 @@ import com.raytheon.uf.edex.decodertools.bufr.packets.IBUFRDataPacket;
  * May 14, 2014  2536     bclement    moved WMO Header to common, removed TimeTools usage
  * Jun 12, 2014  2061     bsteffen    Generate unique stationid
  * Jul 22, 2014  3392     nabowle     ACARSRecord has Float fields instead of Double
- *
+ * Jul 23, 2014  3410     bclement    location changed to floats
+ * 
  * </pre>
- *
+ * 
  * @author jkorman
  * @version 1.0
  */
@@ -485,8 +486,8 @@ public class ACARSDataAdapter {
 
         if ((lat != null) && (lon != null)) {
             loc = new AircraftObsLocation();
-            loc.setLatitude(lat);
-            loc.setLongitude(lon);
+            loc.setLatitude(lat.floatValue());
+            loc.setLongitude(lon.floatValue());
             loc.setLocation(lat, lon);
 
             // We can pick up the height here for some data. Have to look
@@ -538,8 +539,8 @@ public class ACARSDataAdapter {
 
         if ((lat != null) && (lon != null)) {
             loc = new AircraftObsLocation();
-            loc.setLatitude(lat);
-            loc.setLongitude(lon);
+            loc.setLatitude(lat.floatValue());
+            loc.setLongitude(lon.floatValue());
             loc.setLocation(lat, lon);
             generateStationId(loc);
         }
